feat: add version 2 of workflow export entities
