+++ engine/api/user_schema.go
@@ -4,6 +4,7 @@ import (
 	"context"
 	"encoding/json"
 	"fmt"
+	v1 "github.com/ovh/cds/sdk/exportentities/v1"
 	"net/http"
 	"reflect"
 
@@ -30,7 +31,7 @@ func (api *API) getUserJSONSchema() service.Handler {
 
 		var sch *jsonschema.Schema
 		if filter == "" || filter == "workflow" {
-			sch = ref.ReflectFromType(reflect.TypeOf(exportentities.Workflow{}))
+			sch = ref.ReflectFromType(reflect.TypeOf(v1.Workflow{}))
 			buf, _ := json.Marshal(sch)
 			res.Workflow = string(buf)
 		}
